backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
groq==0.11.0
replicate==1.0.7
httpx==0.27.2
Pillow==10.4.0
pydub==0.25.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
